src/backend/requirements.txt
torch
numpy
pyserial
flask
flask-cors